Criação do UserController e EndPoints do CRUD do User

+++ src/main/java/com/josveronez/desafio_astentask/business/services/UserService.java
@@ -1,7 +1,10 @@
 package com.josveronez.desafio_astentask.business.services;
 
+import com.josveronez.desafio_astentask.business.dto.UserRequestDTO;
+import com.josveronez.desafio_astentask.business.dto.UserResponseDTO;
 import com.josveronez.desafio_astentask.business.exceptions.ConflictException;
 import com.josveronez.desafio_astentask.business.exceptions.ResourceNotFoundException;
+import com.josveronez.desafio_astentask.business.mappers.UserMapper;
 import com.josveronez.desafio_astentask.domain.entities.User;
 import com.josveronez.desafio_astentask.domain.repositories.UserRepository;
 import org.springframework.stereotype.Service;
@@ -16,55 +19,63 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
-    public User save(User user){
-        if (userRepository.findUserByEmail(user.getEmail()).isPresent()){
-            throw new ConflictException("O email " + user.getEmail() + "já está em uso.");
+    public UserResponseDTO save(UserRequestDTO requestDTO){
+        if (userRepository.findUserByEmail(requestDTO.email()).isPresent()){
+            throw new ConflictException("O email " + requestDTO.email() + "já está em uso.");
         }
-        return userRepository.save(user);
+        User user = UserMapper.toEntity(requestDTO);
+        User savedUser = userRepository.save(user);
+        return UserMapper.toResponseDTO(savedUser);
     }
 
-    public User findByEmail(String email){
-        return userRepository.findUserByEmail(email).orElseThrow(
+    public UserResponseDTO findByEmail(String email){
+        User user = userRepository.findUserByEmail(email).orElseThrow(
                 () -> new ResourceNotFoundException("Não encontramos nenhum usuário com esse email.")
         );
+        return UserMapper.toResponseDTO(user);
     }
 
-    public User findById(Long id){
-        return userRepository.findById(id).orElseThrow(
+    public UserResponseDTO findById(Long id){
+        User user = userRepository.findById(id).orElseThrow(
                 () -> new ResourceNotFoundException("Não encontramos nenhum usuário com esse id.")
         );
+        return UserMapper.toResponseDTO(user);
     }
 
-    public List<User> findAll(){
-        return userRepository.findAll();
+    public List<UserResponseDTO> findAll(){
+        return userRepository.findAll()
+                .stream()
+                .map(UserMapper::toResponseDTO)
+                .toList();
     }
 
     // Atualiza usuário + mantém os campos, caso vierem nulos (ex: se atualizar só o email, muda o email e mantem os outros campos.)
-    public User updateById(Long id, User user){
-        User userToUpdate = findById(id);
-        if (user.getName() != null){
-            userToUpdate.setName(user.getName());
+    public UserResponseDTO updateById(Long id, UserRequestDTO request){
+        User userToUpdate = userRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Nenhum usuário encontrado com esse id.")
+        );
+        if (request.name() != null){
+            userToUpdate.setName(request.name());
         }
         // Verifica se o novo email não está em uso.
-        if (user.getEmail() != null && !user.getEmail().equals(userToUpdate.getEmail())){
-            if (userRepository.findUserByEmail(user.getEmail()).isPresent()){
+        if (request.email() != null && !request.email().equals(userToUpdate.getEmail())){
+            if (userRepository.findUserByEmail(request.email()).isPresent()){
                 throw new ConflictException("Este email já está sendo utilizado por outro usuário.");
             }
-            userToUpdate.setEmail(user.getEmail());
+            userToUpdate.setEmail(request.email());
         }
-        if (user.getRole() != null) {
-            userToUpdate.setRole(user.getRole());
+        if (request.role() != null) {
+            userToUpdate.setRole(request.role());
         }
-        return userRepository.save(userToUpdate);
+
+        User updatedUser = userRepository.save(userToUpdate);
+        return UserMapper.toResponseDTO(updatedUser);
     }
 
     public void deleteById(Long id){
+        if (!userRepository.existsById(id)){
+            throw new ResourceNotFoundException("Usuário com ID" + id + "não encontrado.");
+        }
         userRepository.deleteById(id);
     }
-
-
-
-
-
-
 }
